+++ content/themes/carbon/footer.php
@@ -23,7 +23,7 @@
 				</div>
 			</div>
 		</footer>
-		<?php if(isValidSession($_COOKIE['session'])) adminBar(); ?>
+		<?php if($session) adminBar(); ?>
 		<?php getScript('script.js'); ?>
 		<?php getThemeScript('script.js'); ?>
 	</body>