+++ content/themes/carbon/header.php
@@ -11,7 +11,6 @@
 		<meta property="og:url" content="<?php $rs_post->getPostUrl(); ?>">
 		<meta property="og:image" content="<?php echo getMediaSrc(getSetting('site_logo', false)); ?>">
 		<meta property="og:description" content="<?php echo !empty($rs_post->getPostMeta('description', false)) ? $rs_post->getPostMeta('description') : trimWords(str_replace(array("\n", "\r"), '', strip_tags($rs_post->getPostContent(false))), 25, '.'); ?>">
-		
 		<link type="image/x-icon" href="<?php echo getMediaSrc(getSetting('site_icon', false)); ?>" rel="icon">
 		<?php getStylesheet('style.css'); ?>
 		<?php getStylesheet('font-awesome.min.css', '5.12.0'); ?>