+++ index.php
@@ -13,14 +13,11 @@ require_once PATH.INC.'/functions.php';
 // Create a Post object
 $rs_post = new Post;
 
-// Check whether the front end theme has a functions.php file and include it if so
-if(file_exists(PATH.CONT.'/functions.php')) require_once PATH.CONT.'/functions.php';
-
 // Check whether the session cookie is set and the user's session is valid
 if(isset($_COOKIE['session']) && isValidSession($_COOKIE['session'])) {
 	// Fetch the user's data
 	$session = getOnlineUser($_COOKIE['session']);
 }
 
-// Include the theme's index file
-require_once PATH.CONT.'/index.php';
+// Include the theme loader file
+require_once PATH.INC.'/load-theme.php';