+++ init.php
@@ -4,24 +4,12 @@
  * @since 1.3.0[a]
  */
 
-// Minimum supported PHP version
-if(!defined('PHP')) define('PHP', '7.3');
+// Include named constants
+require_once __DIR__.'/includes/constants.php';
 
 // Check whether the server is running the required PHP version
 if(phpversion() < PHP)
 	exit('<p>The minimum version of PHP that is supported by ReallySimpleCMS is '.PHP.'; your server is running on '.phpversion().'. Please upgrade to the minimum required version or higher to use this CMS.</p>');
-
-// Absolute path to the root directory
-if(!defined('PATH')) define('PATH', __DIR__);
-
-// Path to the includes directory
-if(!defined('INC')) define('INC', '/includes');
-
-// Path to the admin directory
-if(!defined('ADMIN')) define('ADMIN', '/admin');
-
-// Path to the content directory
-if(!defined('CONT')) define('CONT', '/content');
 
 // Try to initialize the CMS or run setup if the config file doesn't exist
 if(file_exists(PATH.'/config.php')) {
@@ -58,15 +46,6 @@ if(file_exists(PATH.'/config.php')) {
 	
 	// Include global functions
 	require_once PATH.INC.'/globals.php';
-	
-	// Path to the stylesheets directory
-	if(!defined('STYLES')) define('STYLES', INC.'/css');
-	
-	// Path to the scripts directory
-	if(!defined('SCRIPTS')) define('SCRIPTS', INC.'/js');
-	
-	// Path to the uploads directory
-	if(!defined('UPLOADS')) define('UPLOADS', CONT.'/uploads');
 } else {
 	// Redirect to the setup page
 	header('Location: '.ADMIN.'/setup.php');